Refactor components for Parse objects
Components Card and YesDetail now read in immutable objects from the Redux state, and use the getter method to render their contents. This is because Parse objects are a superset of immutable objects, and so it doesn't make sense to cast them to JS objects -- as immutable objects, they are the same serverside and clientside.

I have mocked up the creation of dummy Parse objects for when the server is not available (i.e., when it is being tested on a device, as I have not yet deployed any server) -- but this creation is not tested, and so it probably doesn't work.